+++ src/commands/usage_delta/tests_json.rs
@@ -114,6 +114,7 @@ Inter-|   Receive                                                |  Transmit
                 sample,
                 json,
                 delta,
+                ..
             } => {
                 assert!(sample);
                 assert!(json);